+++ build_info.py
@@ -56,8 +56,8 @@ RELEASES: tuple[Release, ...] = (
             "can stop predicting the other."),
 )
 
-VERSION = "1.4.1"
-CODENAME = "Tetrad"
+VERSION = "1.5.0"
+CODENAME = "Cohesin"
 
 BUILD = f'Build {VERSION} "{CODENAME}"'